commodities_tracker_backend/requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
redis
psycopg2-binary
psycopg2-pool
Flask-Cors
jsonify
flask_cors

# Flask==2.3.2
# requests==2.31.0
